Adding basic functionality
Created functions to calculate the bolometric correction and the
luminosity. Not sure at this state whether I want the functions to
operate on individual operations (you give it a set of photometric
observations given on the same night, it gives you the luminosity) or on
the entire set of observations (load in a list or array of observations,
if spits out a bolometric lightcurve.) At this point, I'm leaning toward
individual observations, since you can always run it multiple times to
get a lightcurve.