Implemented ranges of validity
The code now pays attention to whether or not the color you are feeding
into the polynomial fit is within the valid range of the fit. If the
color is invalid, then the bolometric correction is None. The
calculation of Lbol has been modified so that if the bolometric
correction is None, then log(Lbol) is also None.