Modified set_coefficients to handle setting range
Rather than writing a separate function to handle setting the range min
and range max, I modified set_coefficients by renaming it set_constants
and having it return a tuple containing a list of coefficients, the min,
and the max of the validity range. set_constants is now the one-stop
shop for defining which polynomial to build.