Function calculate_derivative_term is working
In order to calculate the uncertainty in the bolometric correction,
I need to calculate the derivative of the BC with respect to the color.
Since it's just a polynomial, I just needed to tweak the calculate_term
function from a_i * x^i to i * a_i * x^(i-1) in order to be able to call
it in a loop and calculate the derivative of the polynomial.