Changed the way bad BC corrections are returned
In the original code, I was returning None for values of the color which
were outside the range of validity of my polynomial fits.

This is an awkward thing to return, and in practice it's not easy to
handle the output of the code when it will occasinally return None.

So instead, I've changed the error return value to be -999.
In practice, this is not a possible bolometric luminosity for a
supernova, so there is no possibility of mistaking this for good output,
and it avoids the potential pitfalls of having to return None.